Extract rename functionality to a separate service
