+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Rename/RenameService.cs
@@ -1,0 +1,352 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Razor;
+using Microsoft.AspNetCore.Razor.Language;
+using Microsoft.AspNetCore.Razor.Language.Syntax;
+using Microsoft.AspNetCore.Razor.PooledObjects;
+using Microsoft.CodeAnalysis.Razor.DocumentMapping;
+using Microsoft.CodeAnalysis.Razor.ProjectSystem;
+using Microsoft.CodeAnalysis.Razor.Protocol;
+using Microsoft.CodeAnalysis.Razor.Workspaces;
+using Microsoft.VisualStudio.LanguageServer.Protocol;
+using RazorSyntaxKind = Microsoft.AspNetCore.Razor.Language.SyntaxKind;
+using RazorSyntaxNode = Microsoft.AspNetCore.Razor.Language.Syntax.SyntaxNode;
+
+namespace Microsoft.CodeAnalysis.Razor.Rename;
+
+internal class RenameService(
+    IRazorComponentSearchEngine componentSearchEngine,
+    IProjectCollectionResolver projectCollectionResolver,
+    LanguageServerFeatureOptions languageServerFeatureOptions) : IRenameService
+{
+    private readonly IRazorComponentSearchEngine _componentSearchEngine = componentSearchEngine;
+    private readonly IProjectCollectionResolver _projectCollectionResolver = projectCollectionResolver;
+    private readonly LanguageServerFeatureOptions _languageServerFeatureOptions = languageServerFeatureOptions;
+
+    public async Task<WorkspaceEdit?> TryGetRazorRenameEditsAsync(VersionedDocumentContext documentContext, DocumentPositionInfo positionInfo, string newName, CancellationToken cancellationToken)
+    {
+        // We only support renaming of .razor components, not .cshtml tag helpers
+        if (!FileKinds.IsComponent(documentContext.FileKind))
+        {
+            return null;
+        }
+
+        // If we're in C# then there is no point checking for a component tag, because there won't be one
+        if (positionInfo.LanguageKind == RazorLanguageKind.CSharp)
+        {
+            return null;
+        }
+
+        var codeDocument = await documentContext.GetCodeDocumentAsync(cancellationToken).ConfigureAwait(false);
+
+        var originTagHelpers = await GetOriginTagHelpersAsync(documentContext, positionInfo.HostDocumentIndex, cancellationToken).ConfigureAwait(false);
+        if (originTagHelpers.IsDefaultOrEmpty)
+        {
+            return null;
+        }
+
+        var originComponentDocumentSnapshot = await _componentSearchEngine.TryLocateComponentAsync(documentContext.Snapshot, originTagHelpers.First()).ConfigureAwait(false);
+        if (originComponentDocumentSnapshot is null)
+        {
+            return null;
+        }
+
+        var originComponentDocumentFilePath = originComponentDocumentSnapshot.FilePath.AssumeNotNull();
+        var newPath = MakeNewPath(originComponentDocumentFilePath, newName);
+        if (File.Exists(newPath))
+        {
+            return null;
+        }
+
+        using var _ = ListPool<SumType<TextDocumentEdit, CreateFile, RenameFile, DeleteFile>>.GetPooledObject(out var documentChanges);
+        var fileRename = GetFileRenameForComponent(originComponentDocumentSnapshot, newPath);
+        documentChanges.Add(fileRename);
+        AddEditsForCodeDocument(documentChanges, originTagHelpers, newName, documentContext.Uri, codeDocument);
+
+        var documentSnapshots = GetAllDocumentSnapshots(documentContext);
+
+        foreach (var documentSnapshot in documentSnapshots)
+        {
+            await AddEditsForCodeDocumentAsync(documentChanges, originTagHelpers, newName, documentSnapshot).ConfigureAwait(false);
+        }
+
+        foreach (var documentChange in documentChanges)
+        {
+            if (documentChange.TryGetFirst(out var textDocumentEdit) &&
+                textDocumentEdit.TextDocument.Uri == fileRename.OldUri)
+            {
+                textDocumentEdit.TextDocument.Uri = fileRename.NewUri;
+            }
+        }
+
+        return new WorkspaceEdit
+        {
+            DocumentChanges = documentChanges.ToArray(),
+        };
+    }
+
+    private ImmutableArray<IDocumentSnapshot?> GetAllDocumentSnapshots(DocumentContext skipDocumentContext)
+    {
+        using var documentSnapshots = new PooledArrayBuilder<IDocumentSnapshot?>();
+        using var _ = StringHashSetPool.GetPooledObject(out var documentPaths);
+
+        foreach (var project in _projectCollectionResolver.EnumerateProjects(skipDocumentContext.Snapshot))
+        {
+            foreach (var documentPath in project.DocumentFilePaths)
+            {
+                // We've already added refactoring edits for our document snapshot
+                if (string.Equals(documentPath, skipDocumentContext.FilePath, FilePathComparison.Instance))
+                {
+                    continue;
+                }
+
+                // Don't add duplicates between projects
+                if (!documentPaths.Add(documentPath))
+                {
+                    continue;
+                }
+
+                // Add to the list and add the path to the set
+                if (project.GetDocument(documentPath) is not { } snapshot)
+                {
+                    throw new InvalidOperationException($"{documentPath} in project {project.FilePath} but not retrievable");
+                }
+
+                documentSnapshots.Add(snapshot);
+            }
+        }
+
+        return documentSnapshots.DrainToImmutable();
+    }
+
+    private RenameFile GetFileRenameForComponent(IDocumentSnapshot documentSnapshot, string newPath)
+    {
+        // VS Code in Windows expects path to start with '/'
+        var filePath = documentSnapshot.FilePath.AssumeNotNull();
+        var updatedOldPath = _languageServerFeatureOptions.ReturnCodeActionAndRenamePathsWithPrefixedSlash && !filePath.StartsWith("/")
+            ? '/' + filePath
+            : filePath;
+        var oldUri = new UriBuilder
+        {
+            Path = updatedOldPath,
+            Host = string.Empty,
+            Scheme = Uri.UriSchemeFile,
+        }.Uri;
+
+        // VS Code in Windows expects path to start with '/'
+        var updatedNewPath = _languageServerFeatureOptions.ReturnCodeActionAndRenamePathsWithPrefixedSlash && !newPath.StartsWith("/")
+            ? '/' + newPath
+            : newPath;
+        var newUri = new UriBuilder
+        {
+
+            Path = updatedNewPath,
+            Host = string.Empty,
+            Scheme = Uri.UriSchemeFile,
+        }.Uri;
+
+        return new RenameFile
+        {
+            OldUri = oldUri,
+            NewUri = newUri,
+        };
+    }
+
+    private static string MakeNewPath(string originalPath, string newName)
+    {
+        var newFileName = $"{newName}{Path.GetExtension(originalPath)}";
+        var directoryName = Path.GetDirectoryName(originalPath);
+        Assumes.NotNull(directoryName);
+        var newPath = Path.Combine(directoryName, newFileName);
+        return newPath;
+    }
+
+    private async Task AddEditsForCodeDocumentAsync(
+        List<SumType<TextDocumentEdit, CreateFile, RenameFile, DeleteFile>> documentChanges,
+        ImmutableArray<TagHelperDescriptor> originTagHelpers,
+        string newName,
+        IDocumentSnapshot? documentSnapshot)
+    {
+        if (documentSnapshot is null)
+        {
+            return;
+        }
+
+        var codeDocument = await documentSnapshot.GetGeneratedOutputAsync().ConfigureAwait(false);
+        if (codeDocument.IsUnsupported())
+        {
+            return;
+        }
+
+        if (!FileKinds.IsComponent(codeDocument.GetFileKind()))
+        {
+            return;
+        }
+
+        // VS Code in Windows expects path to start with '/'
+        var filePath = documentSnapshot.FilePath.AssumeNotNull();
+        var updatedPath = _languageServerFeatureOptions.ReturnCodeActionAndRenamePathsWithPrefixedSlash && !filePath.StartsWith("/")
+            ? "/" + filePath
+            : filePath;
+        var uri = new UriBuilder
+        {
+            Path = updatedPath,
+            Host = string.Empty,
+            Scheme = Uri.UriSchemeFile,
+        }.Uri;
+
+        AddEditsForCodeDocument(documentChanges, originTagHelpers, newName, uri, codeDocument);
+    }
+
+    private static void AddEditsForCodeDocument(
+        List<SumType<TextDocumentEdit, CreateFile, RenameFile, DeleteFile>> documentChanges,
+        ImmutableArray<TagHelperDescriptor> originTagHelpers,
+        string newName,
+        Uri uri,
+        RazorCodeDocument codeDocument)
+    {
+        var documentIdentifier = new OptionalVersionedTextDocumentIdentifier { Uri = uri };
+        var tagHelperElements = codeDocument.GetSyntaxTree().Root
+            .DescendantNodes()
+            .Where(n => n.Kind == RazorSyntaxKind.MarkupTagHelperElement)
+            .OfType<MarkupTagHelperElementSyntax>();
+
+        foreach (var originTagHelper in originTagHelpers)
+        {
+            var editedName = newName;
+            if (originTagHelper.IsComponentFullyQualifiedNameMatch)
+            {
+                // Fully qualified binding, our "new name" needs to be fully qualified.
+                var @namespace = originTagHelper.GetTypeNamespace();
+                if (@namespace == null)
+                {
+                    return;
+                }
+
+                // The origin TagHelper was fully qualified so any fully qualified rename locations we find will need a fully qualified renamed edit.
+                editedName = @namespace + "." + newName;
+            }
+
+            foreach (var node in tagHelperElements)
+            {
+                if (node is MarkupTagHelperElementSyntax { TagHelperInfo.BindingResult: var binding } tagHelperElement &&
+                    BindingContainsTagHelper(originTagHelper, binding))
+                {
+                    documentChanges.Add(new TextDocumentEdit
+                    {
+                        TextDocument = documentIdentifier,
+                        Edits = CreateEditsForMarkupTagHelperElement(tagHelperElement, codeDocument, editedName),
+                    });
+                }
+            }
+        }
+    }
+
+    private static TextEdit[] CreateEditsForMarkupTagHelperElement(MarkupTagHelperElementSyntax element, RazorCodeDocument codeDocument, string newName)
+    {
+        using var _ = ListPool<TextEdit>.GetPooledObject(out var edits);
+
+        edits.Add(VsLspFactory.CreateTextEdit(element.StartTag.Name.GetRange(codeDocument.Source), newName));
+
+        if (element.EndTag is MarkupTagHelperEndTagSyntax endTag)
+        {
+            edits.Add(VsLspFactory.CreateTextEdit(endTag.Name.GetRange(codeDocument.Source), newName));
+        }
+
+        return [.. edits];
+    }
+
+    private static bool BindingContainsTagHelper(TagHelperDescriptor tagHelper, TagHelperBinding potentialBinding) =>
+        potentialBinding.Descriptors.Any(descriptor => descriptor.Equals(tagHelper));
+
+    private static async Task<ImmutableArray<TagHelperDescriptor>> GetOriginTagHelpersAsync(DocumentContext documentContext, int absoluteIndex, CancellationToken cancellationToken)
+    {
+        var owner = await documentContext.GetSyntaxNodeAsync(absoluteIndex, cancellationToken).ConfigureAwait(false);
+        if (owner is null)
+        {
+            Debug.Fail("Owner should never be null.");
+            return default;
+        }
+
+        var node = owner.FirstAncestorOrSelf<RazorSyntaxNode>(n => n.Kind == RazorSyntaxKind.MarkupTagHelperStartTag);
+        if (node is not MarkupTagHelperStartTagSyntax tagHelperStartTag)
+        {
+            return default;
+        }
+
+        // Ensure the rename action was invoked on the component name
+        // instead of a component parameter. This serves as an issue
+        // mitigation till `textDocument/prepareRename` is supported
+        // and we can ensure renames aren't triggered in unsupported
+        // contexts. (https://github.com/dotnet/aspnetcore/issues/26407)
+        if (!tagHelperStartTag.Name.FullSpan.IntersectsWith(absoluteIndex))
+        {
+            return default;
+        }
+
+        if (tagHelperStartTag?.Parent is not MarkupTagHelperElementSyntax { TagHelperInfo.BindingResult: var binding })
+        {
+            return default;
+        }
+
+        // Can only have 1 component TagHelper belonging to an element at a time
+        var primaryTagHelper = binding.Descriptors.FirstOrDefault(static d => d.IsComponentTagHelper);
+        if (primaryTagHelper is null)
+        {
+            return default;
+        }
+
+        using var originTagHelpers = new PooledArrayBuilder<TagHelperDescriptor>();
+        originTagHelpers.Add(primaryTagHelper);
+
+        var tagHelpers = await documentContext.Snapshot.Project.GetTagHelpersAsync(cancellationToken).ConfigureAwait(false);
+        var associatedTagHelper = FindAssociatedTagHelper(primaryTagHelper, tagHelpers);
+        if (associatedTagHelper is null)
+        {
+            Debug.Fail("Components should always have an associated TagHelper.");
+            return default;
+        }
+
+        originTagHelpers.Add(associatedTagHelper);
+
+        return originTagHelpers.DrainToImmutable();
+    }
+
+    private static TagHelperDescriptor? FindAssociatedTagHelper(TagHelperDescriptor tagHelper, ImmutableArray<TagHelperDescriptor> tagHelpers)
+    {
+        var typeName = tagHelper.GetTypeName();
+        var assemblyName = tagHelper.AssemblyName;
+        foreach (var currentTagHelper in tagHelpers)
+        {
+            if (tagHelper == currentTagHelper)
+            {
+                // Same as the primary, we're looking for our other pair.
+                continue;
+            }
+
+            if (typeName != currentTagHelper.GetTypeName())
+            {
+                continue;
+            }
+
+            if (assemblyName != currentTagHelper.AssemblyName)
+            {
+                continue;
+            }
+
+            // Found our associated TagHelper, there should only ever be 1 other associated TagHelper (fully qualified and non-fully qualified).
+            return currentTagHelper;
+        }
+
+        return null;
+    }
+}
