+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Refactoring/RenameEndpointDelegationTest.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Test.Common.Mef;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
+using Microsoft.CodeAnalysis.Razor.Rename;
 using Microsoft.CodeAnalysis.Razor.Workspaces;
 using Microsoft.CodeAnalysis.Testing;
 using Microsoft.CodeAnalysis.Text;
@@ -66,9 +67,10 @@ public class RenameEndpointDelegationTest(ITestOutputHelper testOutput) : Single
 
         var searchEngine = new RazorComponentSearchEngine(projectManager, LoggerFactory);
 
+        var renameService = new RenameService(searchEngine, projectManager, LanguageServerFeatureOptions);
+
         var endpoint = new RenameEndpoint(
-            searchEngine,
-            projectManager,
+            renameService,
             LanguageServerFeatureOptions,
             DocumentMappingService,
             EditMappingService,
